update refresh journeys endpoint and how task handler calls it
