+++ src/xngin/apiserver/routers/admin_integrations/admin_integrations_api.py
@@ -540,41 +540,3 @@ async def delete_turn_arm_journey_mapping(
     )
     await session.commit()
     return response
-
-
-@router.post(
-    "/integrations/turn/{organization_id}/refresh-journeys",
-    status_code=status.HTTP_204_NO_CONTENT,
-    responses={
-        "404": {
-            "model": dict,
-            "description": "The specified webhook or Turn.io connection was not found.",
-        },
-    },
-)
-async def refetch_journeys_from_turn(
-    organization_id: str,
-    session: Annotated[AsyncSession, Depends(xngin_db_session)],
-    httpx_client: Annotated[httpx2.AsyncClient, Depends(retrying_httpx_dependency)],
-):
-    """
-    This endpoint is used by the refresh task handler to refresh the journeys dictionary by triggering a re-fetch
-    from Turn.io.
-    This is meant to be used ONLY by the Turn.io webhook handler, and not directly by users.
-    """
-    turn_connection = (
-        await session.execute(
-            select(tables.TurnConnection).where(tables.TurnConnection.organization_id == organization_id)
-        )
-    ).scalar_one_or_none()
-
-    if not turn_connection:
-        raise HTTPException(
-            status_code=status.HTTP_404_NOT_FOUND, detail="No Turn.io connection configured for this organization."
-        )
-
-    journeys = await refresh_journeys_dict(turn_connection.get_turn_api_token(), httpx_client)
-    turn_connection.journeys_dict = {journey.name: journey.uuid for journey in journeys}
-    await session.commit()
-
-    return GENERIC_SUCCESS
