+++ src/xngin/apiserver/testing/admin_integrations_api_client.py
@@ -1035,75 +1035,3 @@ class AdminIntegrationsAPIClient:
                 client_exts=client_exts,
             ),
         )
-
-    @overload
-    def refetch_journeys_from_turn(
-        self,
-        *,
-        organization_id: str,
-        raise_if_not_default_status: Literal[True] = True,
-        client_exts: AdminIntegrationsAPIClientExtensions | None = None,
-    ) -> AdminIntegrationsAPIClientResult[Literal[HTTPStatus.NO_CONTENT], Any]: ...
-    @overload
-    def refetch_journeys_from_turn(
-        self,
-        *,
-        organization_id: str,
-        raise_if_not_default_status: Literal[False],
-        client_exts: AdminIntegrationsAPIClientExtensions | None = None,
-    ) -> (
-        AdminIntegrationsAPIClientResult[Literal[HTTPStatus.NO_CONTENT], Any]
-        | AdminIntegrationsAPIClientResult[Literal[HTTPStatus.BAD_REQUEST], HTTPExceptionError]
-        | AdminIntegrationsAPIClientResult[Literal[HTTPStatus.UNAUTHORIZED], HTTPExceptionError]
-        | AdminIntegrationsAPIClientResult[Literal[HTTPStatus.FORBIDDEN], HTTPExceptionError]
-        | AdminIntegrationsAPIClientResult[Literal[HTTPStatus.NOT_FOUND], dict]
-        | AdminIntegrationsAPIClientResult[
-            Literal[HTTPStatus.UNPROCESSABLE_CONTENT], AdminIntegrationsAPIClientHTTPValidationError
-        ]
-    ): ...
-    def refetch_journeys_from_turn(
-        self,
-        *,
-        organization_id: str,
-        raise_if_not_default_status: bool = True,
-        client_exts: AdminIntegrationsAPIClientExtensions | None = None,
-    ) -> (
-        AdminIntegrationsAPIClientResult[Literal[HTTPStatus.NO_CONTENT], Any]
-        | AdminIntegrationsAPIClientResult[Literal[HTTPStatus.BAD_REQUEST], HTTPExceptionError]
-        | AdminIntegrationsAPIClientResult[Literal[HTTPStatus.UNAUTHORIZED], HTTPExceptionError]
-        | AdminIntegrationsAPIClientResult[Literal[HTTPStatus.FORBIDDEN], HTTPExceptionError]
-        | AdminIntegrationsAPIClientResult[Literal[HTTPStatus.NOT_FOUND], dict]
-        | AdminIntegrationsAPIClientResult[
-            Literal[HTTPStatus.UNPROCESSABLE_CONTENT], AdminIntegrationsAPIClientHTTPValidationError
-        ]
-    ):
-        return cast(
-            (
-                AdminIntegrationsAPIClientResult[Literal[HTTPStatus.NO_CONTENT], Any]
-                | AdminIntegrationsAPIClientResult[Literal[HTTPStatus.BAD_REQUEST], HTTPExceptionError]
-                | AdminIntegrationsAPIClientResult[Literal[HTTPStatus.UNAUTHORIZED], HTTPExceptionError]
-                | AdminIntegrationsAPIClientResult[Literal[HTTPStatus.FORBIDDEN], HTTPExceptionError]
-                | AdminIntegrationsAPIClientResult[Literal[HTTPStatus.NOT_FOUND], dict]
-                | AdminIntegrationsAPIClientResult[
-                    Literal[HTTPStatus.UNPROCESSABLE_CONTENT], AdminIntegrationsAPIClientHTTPValidationError
-                ]
-            ),
-            self._route_handler(
-                path="/v1/m/integrations/turn/{organization_id}/refresh-journeys",
-                method=HTTPMethod.POST,
-                default_status=HTTPStatus.NO_CONTENT,
-                models={
-                    HTTPStatus.NO_CONTENT: Any,
-                    HTTPStatus.BAD_REQUEST: HTTPExceptionError,
-                    HTTPStatus.UNAUTHORIZED: HTTPExceptionError,
-                    HTTPStatus.FORBIDDEN: HTTPExceptionError,
-                    HTTPStatus.NOT_FOUND: dict,
-                    HTTPStatus.UNPROCESSABLE_CONTENT: AdminIntegrationsAPIClientHTTPValidationError,
-                },
-                path_params={
-                    "organization_id": organization_id,
-                },
-                raise_if_not_default_status=raise_if_not_default_status,
-                client_exts=client_exts,
-            ),
-        )
